+++ src/test/java/methods/Transfer.java
@@ -7,7 +7,7 @@ import specs.RequestSpecs;
 import specs.ResponseSpecs;
 
 public class Transfer {
-    private static void transferMoney(String userAuthorization, int senderId, int receiverId, double amount, ResponseSpecification response) {
+    private static void transferMoney(String userAuthorization, long senderId, long receiverId, double amount, ResponseSpecification response) {
         TransferRequest transferRequest = TransferRequest.builder()
                 .senderAccountId(senderId)
                 .receiverAccountId(receiverId)
@@ -20,11 +20,11 @@ public class Transfer {
                 .post(transferRequest);
     }
 
-    public static void returnOK(String userAuthorization, int senderId, int receiverId, double amount) {
+    public static void returnOK(String userAuthorization, long senderId, long receiverId, double amount) {
         transferMoney(userAuthorization, senderId, receiverId, amount, ResponseSpecs.requestReturnsOK());
     }
 
-    public static void returnBadRequest(String userAuthorization, int senderId, int receiverId, double amount) {
+    public static void returnBadRequest(String userAuthorization, long senderId, long receiverId, double amount) {
         transferMoney(userAuthorization, senderId, receiverId, amount, ResponseSpecs.requestReturnsBadRequest());
     }
 }
